+++ src/app/core/services/error-handling/error-handling.ts
@@ -7,6 +7,7 @@ export enum ErrorCodes {
   E200 = 'ok',
   E400 = 'bad_request',
   E401 = 'unauthorized',
+  E401_1 = 'invalid_refresh_token',
   E403 = 'forbidden',
   E404 = 'not_found',
   E422 = 'unprocessable_entity',
@@ -25,6 +26,7 @@ export class ErrorHandlerService {
     E400_DEFAULT_BAD_REQUEST: 'Die Daten sind nicht korrekt.',
     E400_LOGIN_BAD_REQUEST: 'E-Mail oder Passwort sind nicht korrekt.',
     E401_DEFAULT_UNAUTHORIZED: 'Die Anfrage kann ohne Authentifizierung nicht durchgeführt werden.',
+    E401_DEFAULT_INVALID_REFRESH_TOKEN: 'Du warst lange nicht aktiv auf dieser Seite. Du wurdest deshalb ausgeloggt. Bitte logge dich erneut ein.',
     E403_DEFAULT_FORBIDDEN: 'Es fehlt die Berechtigung Daten zu lesen oder zu bearbeiten.',
     E403_OFFER_FORBIDDEN: 'Es fehlen die Rechte um diesen Kurs zu bearbeiten.',
     E404_DEFAULT_NOT_FOUND: 'Die angeforderten Daten konnten nicht gefunden werden.',
@@ -97,6 +99,9 @@ export class ErrorHandlerService {
     // 401 - (Unauthorized) - Zugriff auf Serverbereiche die nicht authorisiert sind
     this.defaultErrorMap.set(ErrorCodes.E401, this.ERROR_MESSAGES.E401_DEFAULT_UNAUTHORIZED);
 
+    // 401 - (Unauthorized-Invalid RefreshToken) - Das Token konnte nicht mehr upgedated werden.
+    this.defaultErrorMap.set(ErrorCodes.E401_1, this.ERROR_MESSAGES.E401_DEFAULT_INVALID_REFRESH_TOKEN);
+
     // 403 - (Forbidden)
     this.defaultErrorMap.set(ErrorCodes.E403, this.ERROR_MESSAGES.E403_DEFAULT_FORBIDDEN);
 
@@ -123,6 +128,5 @@ export class ErrorHandlerService {
      // Für offer
     this.offerErrorMap.set(ErrorCodes.E403, this.ERROR_MESSAGES.E403_OFFER_FORBIDDEN);
     this.offerErrorMap.set(ErrorCodes.E404, this.ERROR_MESSAGES.E404_OFFER_NOT_FOUND);
-   
   }
 }