+++ Changelog.md
@@ -2,6 +2,10 @@
 
 ## [Not Released v0.10.0]
 
+## 2022-03-01/ml
+### changed
+- Interceptor/API/AuthService check if 401 error occurs to refresh token
+
 ## 2022-02-23/avz
 ### added
 - Loading spinner on the Detail Page